API: Move the context from the handlers into the router.
